requirements.txt
requests
tqdm
opencv-python
Pillow>=9.0.0
natsort
numpy>=1.21.0
tenacity>=8.0.0
img2pdf>=0.4.4
# Enhancement dependencies (temporarily disabled due to compatibility issues)
# torch==2.0.1
# torchvision==0.15.2
# basicsr==1.4.2
# realesrgan==0.3.0
